test(e2e): Rename test packages
