+++ .github/workflows/enos-run.yml
@@ -78,8 +78,8 @@ jobs:
         include:
           - filter: 'e2e_aws builder:crt'
           - filter: 'e2e_database'
-          - filter: 'e2e_static builder:crt'
-          - filter: 'e2e_static_with_vault builder:crt'
+          - filter: 'e2e_base builder:crt'
+          - filter: 'e2e_base_with_vault builder:crt'
           # - filter: 'e2e_ui builder:crt' # Don't run UI tests yet. takes too long.
     runs-on: ${{ fromJSON(vars.RUNNER) }}
     env:
@@ -182,7 +182,7 @@ jobs:
         run: |
           wget https://releases.hashicorp.com/vault/1.12.2/vault_1.12.2_linux_amd64.zip -O /tmp/test-deps/vault.zip
       - name: Install Vault CLI
-        if: matrix.filter == 'e2e_static_with_vault builder:crt' || matrix.filter == 'e2e_database' || matrix.filter == 'e2e_ui builder:crt' || matrix.filter == 'e2e_docker builder:crt'
+        if: matrix.filter == 'e2e_base_with_vault builder:crt' || matrix.filter == 'e2e_database' || matrix.filter == 'e2e_ui builder:crt' || matrix.filter == 'e2e_docker builder:crt'
         run: |
           unzip /tmp/test-deps/vault.zip -d /usr/local/bin
       - name: Add hosts to /etc/hosts
