+++ enos/enos-scenario-e2e-base.hcl
@@ -1,7 +1,7 @@
 # Copyright (c) HashiCorp, Inc.
 # SPDX-License-Identifier: MPL-2.0
 
-scenario "e2e_static" {
+scenario "e2e_base" {
   terraform_cli = terraform_cli.default
   terraform     = terraform.default
   providers = [
@@ -121,7 +121,7 @@ scenario "e2e_static" {
     ]
 
     variables {
-      test_package             = "github.com/hashicorp/boundary/testing/internal/e2e/tests/static"
+      test_package             = "github.com/hashicorp/boundary/testing/internal/e2e/tests/base"
       debug_no_run             = var.e2e_debug_no_run
       alb_boundary_api_addr    = step.create_boundary_cluster.alb_boundary_api_addr
       auth_method_id           = step.create_boundary_cluster.auth_method_id
