+++ testing/internal/e2e/tests/base/connect_ssh_test.go
@@ -1,14 +1,12 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package static_test
+package base_test
 
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"os"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/boundary/api/targets"
@@ -18,9 +16,10 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// TestCliConnectTargetWithAuthzToken uses the boundary cli to connect to a target using the
-// `authz_token` option
-func TestCliConnectTargetWithAuthzToken(t *testing.T) {
+// TestCliConnectTargetWithSsh uses the boundary cli to create a credential using boundary's
+// built-in credential store. The test attaches that credential to a target and attempts to connect
+// to that target using those credentials.
+func TestCliConnectTargetWithSsh(t *testing.T) {
 	e2e.MaybeSkipTest(t)
 	c, err := loadTestConfig()
 	require.NoError(t, err)
@@ -67,38 +66,16 @@ func TestCliConnectTargetWithAuthzToken(t *testing.T) {
 	require.Equal(t, string(k), retrievedKey)
 	t.Log("Successfully retrieved credentials for target")
 
-	// Get auth token for session
-	newAuthToken := newSessionAuthorizationResult.Item.AuthorizationToken
-
-	// Create key file
-	retrievedKeyPath := fmt.Sprintf("%s/%s", t.TempDir(), "target_private_key.pem")
-	f, err := os.Create(retrievedKeyPath)
-	require.NoError(t, err)
-	_, err = f.WriteString(retrievedKey)
-	require.NoError(t, err)
-	err = os.Chmod(retrievedKeyPath, 0o400)
-	require.NoError(t, err)
-
-	// Connect to target and print host's IP address using retrieved credentials
+	// Connect to target using ssh option
 	output = e2e.RunCommand(ctx, "boundary",
 		e2e.WithArgs(
-			"connect",
-			"-authz-token", newAuthToken,
-			"-exec", "/usr/bin/ssh", "--",
-			"-l", retrievedUser,
-			"-i", retrievedKeyPath,
+			"connect", "ssh",
+			"-target-id", newTargetId, "--",
 			"-o", "UserKnownHostsFile=/dev/null",
 			"-o", "StrictHostKeyChecking=no",
 			"-o", "IdentitiesOnly=yes", // forces the use of the provided key
-			"-p", "{{boundary.port}}", // this is provided by boundary
-			"{{boundary.ip}}",
-			"hostname", "-i",
 		),
 	)
 	require.NoError(t, output.Err, string(output.Stderr))
-
-	parts := strings.Fields(string(output.Stdout))
-	hostIp := parts[len(parts)-1]
-	require.Equal(t, c.TargetIp, hostIp, "SSH session did not return expected output")
 	t.Log("Successfully connected to target")
 }
