+++ testing/internal/e2e/tests/base/session_end_delete_target_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package static_test
+package base_test
 
 import (
 	"context"
@@ -15,10 +15,13 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// TestCliSessionCancelAdmin uses the boundary cli to start and then cancel a session
-func TestCliSessionCancelAdmin(t *testing.T) {
+// TestCliSessionEndWhenTargetIsDeleted tests that an active session is canceled when the respective
+// target for the session is deleted.
+func TestCliSessionEndWhenTargetIsDeleted(t *testing.T) {
 	e2e.MaybeSkipTest(t)
 	c, err := loadTestConfig()
+	require.NoError(t, err)
+	bc, err := boundary.LoadConfig()
 	require.NoError(t, err)
 
 	ctx := context.Background()
@@ -37,15 +40,38 @@ func TestCliSessionCancelAdmin(t *testing.T) {
 	boundary.AddHostToHostSetCli(t, ctx, newHostSetId, newHostId)
 	newTargetId := boundary.CreateNewTargetCli(t, ctx, newProjectId, c.TargetPort)
 	boundary.AddHostSourceToTargetCli(t, ctx, newTargetId, newHostSetId)
+	acctName := "e2e-account"
+	newAccountId, acctPassword := boundary.CreateNewAccountCli(t, ctx, bc.AuthMethodId, acctName)
+	t.Cleanup(func() {
+		boundary.AuthenticateAdminCli(t, context.Background())
+		output := e2e.RunCommand(ctx, "boundary",
+			e2e.WithArgs("accounts", "delete", "-id", newAccountId),
+		)
+		require.NoError(t, output.Err, string(output.Stderr))
+	})
+	newUserId := boundary.CreateNewUserCli(t, ctx, "global")
+	t.Cleanup(func() {
+		boundary.AuthenticateAdminCli(t, context.Background())
+		output := e2e.RunCommand(ctx, "boundary",
+			e2e.WithArgs("users", "delete", "-id", newUserId),
+		)
+		require.NoError(t, output.Err, string(output.Stderr))
+	})
+	boundary.SetAccountToUserCli(t, ctx, newUserId, newAccountId)
+	newRoleId := boundary.CreateNewRoleCli(t, ctx, newProjectId)
+	boundary.AddGrantToRoleCli(t, ctx, newRoleId, "id=*;type=target;actions=authorize-session")
+	boundary.AddPrincipalToRoleCli(t, ctx, newRoleId, newUserId)
 
 	// Connect to target to create a session
 	ctxCancel, cancel := context.WithCancel(context.Background())
 	errChan := make(chan *e2e.CommandResult)
 	go func() {
-		t.Log("Starting session...")
+		token := boundary.GetAuthenticationTokenCli(t, ctx, acctName, acctPassword)
+		t.Log("Starting session as user...")
 		errChan <- e2e.RunCommand(ctxCancel, "boundary",
 			e2e.WithArgs(
 				"connect",
+				"-token", "env://E2E_AUTH_TOKEN",
 				"-target-id", newTargetId,
 				"-exec", "/usr/bin/ssh", "--",
 				"-l", c.TargetSshUser,
@@ -57,6 +83,7 @@ func TestCliSessionCancelAdmin(t *testing.T) {
 				"{{boundary.ip}}",
 				"hostname -i; sleep 60",
 			),
+			e2e.WithEnv("E2E_AUTH_TOKEN", token),
 		)
 	}()
 	t.Cleanup(cancel)
@@ -65,15 +92,13 @@ func TestCliSessionCancelAdmin(t *testing.T) {
 	assert.Equal(t, newTargetId, s.TargetId)
 	assert.Equal(t, newHostId, s.HostId)
 
-	// Cancel session
-	t.Log("Canceling session...")
-	output := e2e.RunCommand(ctx, "boundary",
-		e2e.WithArgs("sessions", "cancel", "-id", s.Id),
-	)
+	// Delete Target
+	t.Log("Deleting target...")
+	output := e2e.RunCommand(ctx, "boundary", e2e.WithArgs("targets", "delete", "-id", newTargetId))
 	require.NoError(t, output.Err, string(output.Stderr))
-	boundary.WaitForSessionStatusCli(t, ctx, s.Id, session.StatusTerminated.String())
 
-	// Check output from session
+	// Check if session has terminated
+	t.Log("Waiting for session to be canceling/terminated...")
 	select {
 	case output := <-errChan:
 		// `boundary connect` returns a 255 when cancelled
@@ -82,5 +107,6 @@ func TestCliSessionCancelAdmin(t *testing.T) {
 		t.Fatal("Timed out waiting for session command to exit")
 	}
 
-	t.Log("Successfully cancelled session")
+	boundary.WaitForSessionStatusCli(t, ctx, s.Id, session.StatusTerminated.String())
+	t.Log("Session successfully ended after target was deleted")
 }
