+++ testing/internal/e2e/tests/base/testdata/credential.json
