+++ testing/internal/e2e/tests/base_with_vault/testdata/boundary-controller-policy.hcl
